Support for .NET Standard 2.0 / .NET 4.6.1
Add support for .NET Standard 2.0 (in all the core libraries) and
.NET 4.6.1 (in the OMS and test projects).

Summary of changes:

- Update all projects to [SDK style][1]; set GenerateAssemblyInfo to false
  to [avoid having to touch existing AssemblyInfo files][2]
- Use <PackageReference> instead of packages.config
- Update .sln file to remove unnecessary build targets (assume AnyCPU)
- Upgrade to newest NUnit
- Modify path prefix in OrcaMDF.RawCore.Tests.BaseFixture
- Move TestHelper out of core framework and include it as a linked file in
  all the test projects (this is needed because the code used here is
  [not compatible with .NET Standard][3])

[1]: https://docs.microsoft.com/en-us/dotnet/core/tools/csproj
[2]: https://github.com/dotnet/runtime/issues/26290
[3]: https://stackoverflow.com/questions/42138418/equivalent-to-assemblyinfo-in-dotnet-core-csproj
