+++ .gitignore
@@ -6,3 +6,4 @@ _ReSharper.OrcaMDF
 DBTests
 *.linq
 OrcaMDF.sln.cache
+.vs/*
