--- src/OrcaMDF.Core.Tests/packages.config
@@ -1,4 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="NUnit" version="2.6.2" targetFramework="net40" />
-</packages>
