+++ src/OrcaMDF.RawCore.Tests/BaseFixture.cs
@@ -2,9 +2,11 @@
 {
 	public class BaseFixture
 	{
-		protected const string AW2012Path = @"TestDatabases\AWLT2012.mdf";
-		protected const string AW2008R2Path = @"TestDatabases\AWLT2008R2.mdf";
-		protected const string AW2008Path = @"TestDatabases\AWLT2008.mdf";
-		protected const string AW2005Path = @"TestDatabases\AWLT2005.mdf";
+		private const string Prefix = @"src\OrcaMDF.RawCore.Tests\TestDatabases\";
+
+		protected const string AW2012Path = Prefix + @"AWLT2012.mdf";
+		protected const string AW2008R2Path = Prefix + @"AWLT2008R2.mdf";
+		protected const string AW2008Path = Prefix + @"AWLT2008.mdf";
+		protected const string AW2005Path = Prefix + @"AWLT2005.mdf";
 	}
 }
